# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/pods": {
            "get": {
                "description": "Retrieves a list of pods from the Kubernetes cluster. You can filter results by namespace or paginate the response using the limit and continue parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pods"
                ],
                "summary": "List pods",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Maximum number of pods to retrieve",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination token for fetching more pods",
                        "name": "continue",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Namespace to filter pods by",
                        "name": "namespace",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of pods",
                        "schema": {
                            "$ref": "#/definitions/controller.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Interval error",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new pod in the Kubernetes cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pods"
                ],
                "summary": "Create a new pod",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Pod request body",
                        "name": "pod",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PodsCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created the pod",
                        "schema": {
                            "$ref": "#/definitions/controller.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request or invalid data",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Interval error",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            }
        },
        "/pods/{id}": {
            "get": {
                "description": "Retrieves a pod from the Kubernetes cluster by its name or UID, optionally filtered by namespace.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pods"
                ],
                "summary": "Get a pod by name or UID",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace to filter the pod by",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name or UID of the pod",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the requested pod",
                        "schema": {
                            "$ref": "#/definitions/controller.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Interval error",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update specific fields of an existing pod in the Kubernetes cluster. The following fields are changeable:\n- containers.image\n- initContainers.image\n- tolerations (only additions)\n- activeDeadlineSeconds\n- terminationGracePeriodSeconds",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pods"
                ],
                "summary": "Update an existing pod",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Pod update request body",
                        "name": "pod",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PodsUpdateRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Namespace to filter the pod by",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name or UID of the pod",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pod successfully updated",
                        "schema": {
                            "$ref": "#/definitions/controller.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request or invalid input data",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    },
                    "500": {
                        "description": "Interval error",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a pod from the Kubernetes cluster by its name or UID, optionally filtered by namespace.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pods"
                ],
                "summary": "Delete a pod by name or UID",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Namespace to filter the pod by",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name or UID of the pod",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Interval error",
                        "schema": {
                            "$ref": "#/definitions/controller.FailureResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.FailureResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "controller.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "model.Container": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "command": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "env": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.EnvVar"
                    }
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ContainerPort"
                    }
                },
                "stdin": {
                    "type": "boolean"
                },
                "stdinOnce": {
                    "type": "boolean"
                },
                "terminationMessagePath": {
                    "type": "string"
                },
                "tty": {
                    "type": "boolean"
                },
                "workingDir": {
                    "type": "string"
                }
            }
        },
        "model.ContainerPort": {
            "type": "object",
            "properties": {
                "containerPort": {
                    "type": "integer"
                },
                "hostIP": {
                    "type": "string"
                },
                "hostPort": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                }
            }
        },
        "model.ContainerRequest": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                },
                "name": {
                    "description": "cannot changable",
                    "type": "string"
                }
            }
        },
        "model.CreateOptions": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "dryRun": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fieldManager": {
                    "type": "string"
                },
                "fieldValidation": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                }
            }
        },
...
        "model.EnvVar": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "model.ObjectMeta": {
            "type": "object",
            "properties": {
                "annotations": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "creationTimestamp": {
                    "type": "string"
                },
                "deletionGracePeriodSeconds": {
                    "type": "integer"
                },
                "deletionTimestamp": {
                    "type": "string"
                },
                "finalizers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "generateName": {
                    "type": "string"
                },
                "generation": {
                    "type": "integer"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "ownerReferences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OwnerReference"
                    }
                },
                "resourceVersion": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "model.OwnerReference": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "blockOwnerDeletion": {
                    "type": "boolean"
                },
                "controller": {
                    "type": "boolean"
                },
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Pod": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/model.ObjectMeta"
                },
                "spec": {
                    "$ref": "#/definitions/model.PodSpec"
                },
                "status": {
                    "$ref": "#/definitions/model.PodStatus"
                }
            }
        },
        "model.PodCondition": {
            "type": "object",
            "properties": {
                "lastProbeTime": {
                    "type": "string"
                },
                "lastTransitionTime": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.PodSpec": {
            "type": "object",
            "properties": {
                "activeDeadlineSeconds": {
                    "type": "integer"
                },
                "containers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Container"
                    }
                },
                "initContainers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Container"
                    }
                },
                "terminationGracePeriodSeconds": {
                    "type": "integer"
                },
                "tolerations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Toleration"
                    }
                },
                "volumes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Volume"
                    }
                }
            }
        },
        "model.PodStatus": {
            "type": "object",
            "properties": {
                "conditions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PodCondition"
                    }
                },
                "hostIP": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "nominatedNodeName": {
                    "type": "string"
                },
                "phase": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "model.PodUpdate": {
            "type": "object",
            "properties": {
                "spec": {
                    "$ref": "#/definitions/model.SpecRequest"
                }
            }
        },
        "model.PodsCreateRequest": {
            "type": "object",
            "properties": {
                "opts": {
                    "$ref": "#/definitions/model.CreateOptions"
                },
                "pod": {
                    "$ref": "#/definitions/model.Pod"
                }
            }
        },
        "model.PodsUpdateRequest": {
            "type": "object",
            "properties": {
                "opts": {
                    "$ref": "#/definitions/model.UpdateOptions"
                },
                "pod": {
                    "$ref": "#/definitions/model.PodUpdate"
                }
            }
        },
        "model.SpecRequest": {
            "type": "object",
            "properties": {
                "activeDeadlineSeconds": {
                    "type": "integer"
                },
                "containers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ContainerRequest"
                    }
                },
                "initContainers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ContainerRequest"
                    }
                },
                "terminationGracePeriodSeconds": {
                    "type": "integer"
                },
                "tolerations": {
                    "description": "allow it to be set to 1 if it was previously negative",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Toleration"
                    }
                }
            }
        },
        "model.Toleration": {
            "type": "object",
            "properties": {
                "effect": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "operator": {
                    "type": "string"
                },
                "tolerationSeconds": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "model.UpdateOptions": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "dryRun": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fieldManager": {
                    "type": "string"
                },
                "fieldValidation": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                }
            }
        },
...
        "model.Volume": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        }
    }
}